telegram_bot/requirements.txt
aiogram==3.15.0
pydantic>=2.4.1
fastapi
uvicorn
python-dotenv
aiohttp
